+++ src/app/(public)/login/LoginForm.tsx
@@ -1,0 +1,91 @@
+"use client";
+import { useForm } from "react-hook-form";
+import { zodResolver } from "@hookform/resolvers/zod";
+import * as z from "zod";
+import { Button } from "@/components/ui/button";
+import {
+  Form,
+  FormControl,
+  FormField,
+  FormItem,
+  FormLabel,
+  FormMessage,
+} from "@/components/ui/form";
+import { Input } from "@/components/ui/input";
+import { login } from "../../../lib/login";
+import { toast } from "sonner";
+
+const USERNAME_VALIDATION_MESSAGE =
+  "username must be between 4 ~ 31 characters, and only consists of lowercase letters, 0-9, -, and _";
+const PASSWORD_VALIDATION_MESSAGE =
+  "password must be between 6 ~ 255 characters";
+
+const LoginSchema = z.object({
+  username: z
+    .string()
+    .min(4, USERNAME_VALIDATION_MESSAGE)
+    .max(31, USERNAME_VALIDATION_MESSAGE),
+  password: z
+    .string()
+    .min(6, PASSWORD_VALIDATION_MESSAGE)
+    .max(255, PASSWORD_VALIDATION_MESSAGE),
+});
+
+type FormValues = z.infer<typeof LoginSchema>;
+
+export const LoginForm = () => {
+  const form = useForm<FormValues>({
+    resolver: zodResolver(LoginSchema),
+    defaultValues: {
+      username: "",
+      password: "",
+    },
+  });
+
+  return (
+    <>
+      <Form {...form}>
+        <form
+          action={async (formData) => {
+            const response = await login(formData);
+            if (response.error) {
+              console.error(response.error);
+              toast.error(response.error);
+            }
+          }}
+          className="space-y-6"
+        >
+          <FormField
+            control={form.control}
+            name="username"
+            render={({ field }) => (
+              <FormItem>
+                <FormLabel className="text-pure-white">Username</FormLabel>
+                <FormControl>
+                  <Input {...field} />
+                </FormControl>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+          <FormField
+            control={form.control}
+            name="password"
+            render={({ field }) => (
+              <FormItem>
+                <FormLabel className="text-pure-white">Password</FormLabel>
+                <FormControl>
+                  <Input type="password" {...field} />
+                </FormControl>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+          <Button className="text-pure-white" type="submit">
+            Log in
+          </Button>
+        </form>
+      </Form>
+    </>
+  );
+};
